+++ lookup/create_lookup_table.py
@@ -10,11 +10,43 @@ anything that can be accessed with the api directly
 """
 import json
 import time
+import requests
 import xml.etree.ElementTree as ET
 
-import id_lookup
 import pandas as pd
 import utils
+
+
+def group_id_to_name(id_list, group_name):
+    """
+    asks the api for names to ids then collects into a dict
+    for one group at a time
+
+    id_list : list of ids given as ints
+    group_name : string (one of [artist, publisher,
+                                 designer, game])
+
+    returns : dictionary
+
+    eg:
+    group_id_to_name([100, 150], "publisher")
+    """
+    id_name_dict = {}
+    id_string = [str(x) for x in id_list]
+    id_string = ",".join(id_string)
+    url = "https://www.boardgamegeek.com/xmlapi/boardgame"
+    if group_name != "game":
+        url = url + group_name
+    url += "/{}".format(id_string)
+    print(url)
+    resp = requests.get(url)
+    tree = ET.fromstring(resp.content)
+    for idx, val in enumerate(list(tree)):
+        name = val.find("name").text
+        id = id_list[idx]
+        id_name_dict[id] = name
+    return id_name_dict
+
 
 if __name__ == "__main__":
 
@@ -41,7 +73,7 @@ if __name__ == "__main__":
 
     for id_chunk in chunked_list:
         try:
-            id_and_name = id_lookup.group_id_to_name(id_chunk, type_of_ids)
+            id_and_name = group_id_to_name(id_chunk, type_of_ids)
             ids_and_names.update(id_and_name)
         except UnboundLocalError as error:
             print(error)
