--- lookup/id_lookup.py
@@ -1,56 +1,0 @@
-"""
-working with the bgg api
-eg:
-https://www.boardgamegeek.com/xmlapi/boardgame/13,150
-"""
-
-import xml.etree.ElementTree as ET
-
-import requests
-
-
-def group_id_to_name(id_list, group_name):
-    """
-    asks the api for names to ids then collects into a dict
-    for one group at a time
-
-    id_list : list of ids given as ints
-    group_name : string (one of [artist, publisher,
-                                 designer, game])
-
-    returns : dictionary
-
-    eg:
-    group_id_to_name([100, 150], "publisher")
-    """
-    id_name_dict = {}
-    id_string = [str(x) for x in id_list]
-    id_string = ",".join(id_string)
-    url = "https://www.boardgamegeek.com/xmlapi/boardgame"
-    if group_name != "game":
-        url = url + group_name
-    url += "/{}".format(id_string)
-    print(url)
-    resp = requests.get(url)
-    tree = ET.fromstring(resp.content)
-    for idx, val in enumerate(list(tree)):
-        name = val.find("name").text
-        id = id_list[idx]
-        id_name_dict[id] = name
-    return id_name_dict
-
-
-if __name__ == "__main__":
-    # url = "https://www.boardgamegeek.com/xmlapi/boardgame/13"
-    # resp = requests.get(url)
-    # tree = ET.fromstring(resp.content)
-
-    pub_list = [13, 150]
-    pubs = group_id_to_name(pub_list, "game")
-    print(pubs)
-
-    # import pdb
-
-    # pdb.set_trace()
-
-    # print(tree)
